Updated the english translation. (99%)
So, now 99% of the file is translated.
There are some FAQ to translate, but I don't think it is necessary since it applies for the chinese users that use some special cards with the app, which isnt available internationally. If there's need to do it, I will do it.

If you found any errors (in app) let me know.